Breadcrumbs - split into original Breadcrumbs + BreadcrumbsBar with NotificationToggle
