+++ app/views/layouts/_breadcrumbs.html.haml
@@ -1,5 +1,5 @@
 - right_cell_text ||= @right_cell_text
-= react('Breadcrumbs',
+= react('BreadcrumbsBar',
         :items => (data_for_breadcrumbs({:right_cell_text => right_cell_text}) if respond_to?(:data_for_breadcrumbs)),
         :controllerName => controller_name,
         :title => @title)
